Rename package product
