+++ src/test/java/Product_PM_01_AddNewProduct/VerifyUserFailedAddProductWhenAllFieldsAreBlank.java
@@ -1,4 +1,4 @@
-package ProductCases;
+package Product_PM_01_AddNewProduct;
 
 import example.TestBase;
 import io.qameta.allure.Allure;
